# domain/src/main/java/com/wikiblaz/straviarunning/domain/model/ActivityType.java
package com.wikiblaz.straviarunning.domain.model;

/**
 * Enum representing different types of activities
 */
public enum ActivityType {
    RUN,
    TRAIL_RUN,
    TRACK_RUN,
    TREADMILL,
    RACE,
    WALK,
    HIKE
}