# rest/src/main/resources/api-contract.yaml
openapi: 3.0.3
info:
  title: Stravia Running API
  description: REST API for Stravia Running application
  version: 1.0.0
  contact:
    name: Wikiblaz
    url: https://wikiblaz.com
    email: juablaz@hotmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: activities
    description: Operations related to running activities

paths:
  /api/activities:
    get:
      tags:
        - activities
      summary: Get all activities
      description: Retrieves a list of all running activities
      operationId: getAllActivities
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
    post:
      tags:
        - activities
      summary: Create a new activity
      description: Creates a new running activity
      operationId: createActivity
      requestBody:
        description: Activity object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Invalid input

  /api/activities/{id}:
    get:
      tags:
        - activities
      summary: Get activity by ID
      description: Retrieves a specific activity by its ID
      operationId: getActivityById
      parameters:
        - name: id
          in: path
          description: ID of the activity to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '404':
          description: Activity not found
    put:
      tags:
        - activities
      summary: Update an activity
      description: Updates an existing activity
      operationId: updateActivity
      parameters:
        - name: id
          in: path
          description: ID of the activity to update
          required: true
          schema:
            type: string
      requestBody:
        description: Updated activity object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityRequest'
      responses:
        '200':
          description: Activity updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Invalid input
        '404':
          description: Activity not found
    delete:
      tags:
        - activities
      summary: Delete an activity
      description: Deletes an activity by its ID
      operationId: deleteActivity
      parameters:
        - name: id
          in: path
          description: ID of the activity to delete
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Activity deleted successfully
        '404':
          description: Activity not found

components:
  schemas:
    ActivityType:
      type: string
      enum:
        - RUN
        - TRAIL_RUN
        - TRACK_RUN
        - TREADMILL
        - RACE
        - WALK
        - HIKE
      
    ActivityRequest:
      type: object
      required:
        - name
        - distance
        - duration
        - startDate
        - type
      properties:
        name:
          type: string
          example: "Morning Run"
        description:
          type: string
          example: "Easy morning run around the park"
        distance:
          type: number
          format: double
          description: Distance in meters
          example: 5000
        duration:
          type: integer
          format: int64
          description: Duration in seconds
          example: 1800
        startDate:
          type: string
          format: date-time
          example: "2023-05-15T08:30:00Z"
        endDate:
          type: string
          format: date-time
          example: "2023-05-15T09:00:00Z"
        elevationGain:
          type: number
          format: double
          description: Elevation gain in meters
          example: 50
        type:
          $ref: '#/components/schemas/ActivityType'
          
    Activity:
      allOf:
        - $ref: '#/components/schemas/ActivityRequest'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
              example: "123e4567-e89b-12d3-a456-426614174000"